requirements.txt
#-e git+https://github.com/RasaHQ/rasa_core.git@d63f55c44cd919d2031311201caa1c26c94cb365#egg=rasa_core
#rasa_nlu==0.12.0
gspread==2.0.0
oauth2client==4.1.2
duckling==1.7.3
aniso8601==3.0.0
awscli==1.15.33
beautifulsoup4==4.6.0
boto3==1.7.33
botocore==1.10.33
bs4==0.0.1
certifi==2018.4.16
chardet==3.0.4
click==6.7
colorama==0.3.9
Django==2.0.6
docutils==0.14
Flask==1.0.2
Flask-Cors==3.0.4
Flask-RESTful==0.3.6
gTTS==2.0.0
gTTS-token==1.1.1
idna==2.6
itsdangerous==0.24
Jinja2==2.10
jmespath==0.9.3
MarkupSafe==1.0
mpg123==0.4
mutagen==1.40.0
pyasn1==0.4.3
PyAudio==0.2.11
python-dateutil==2.7.3
python-vlc==3.0.102
pytz==2018.4
PyYAML==3.12
requests==2.18.4
rsa==3.4.2
s3transfer==0.1.13
six==1.11.0
SpeechRecognition==3.8.1
urllib3==1.22
Werkzeug==0.14.1 
